+++ example/asyncGenerate.js
@@ -1,7 +1,7 @@
 import * as THREE from 'three';
 import Stats from 'stats.js';
 import { GUI } from 'three/examples/jsm/libs/lil-gui.module.min.js';
-import { ParallelMeshBVHWorker } from '../src/workers/parallel/ParallelMeshBVHWorker.js';
+import { ParallelMeshBVHWorker } from '../src/workers/ParallelMeshBVHWorker.js';
 import { acceleratedRaycast, AVERAGE, CENTER, MeshBVH, MeshBVHHelper, SAH } from '..';
 
 THREE.Mesh.raycast = acceleratedRaycast;
